% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file_config_ini_new.R
\name{copy_file_config_ini_new}
\alias{copy_file_config_ini_new}
\title{Copy file configuration from outline in working directory}
\usage{
copy_file_config_ini_new(path_to = NULL, overwritefile = TRUE)
}
\arguments{
\item{path_to}{path where the configuration file will be copied}

\item{overwritefile}{logical variable;  if \code{TRUE} will overwrite the file if it already exists}
}
\description{
Copy file configuration from outline in working directory
to search for additional terms in our R Markdown or LaTeX project.
}
\details{
The name of the configuration file has to be "addinsOutline_ini.txt".

Regular expressions are used to perform searches through
the functions of the "stringr" package.

Make a copy of this file to see how new searches are defined.

The following strings serve to distinguish which types of files will be used:

- "rmd": R Markdown files

- "rmddb": Bookdown files

- "tex": LaTeX files
}
\examples{
\dontrun{
copy_file_config_ini_new()
}
}
\seealso{
The function \code{\link{run_addinsOutline_Rmd}()},
\code{\link{run_addinsOutline_Rmd_bookdown}()} and
\code{\link{run_addinsOutline_tex}()}.
}
