% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_addinsOutline_Rmd.R
\name{run_addinsOutline_Rmd}
\alias{run_addinsOutline_Rmd}
\title{Addins Displays Outline of an R Markdown file}
\usage{
run_addinsOutline_Rmd()
}
\value{
RStudio Addins with browserViewer()
}
\description{
Addins Displays Outline of an R Markdown file with
and without child files.
}
\examples{
\dontrun{
run_addinsOutline_Rmd()
}
}
\seealso{
The function \code{\link{run_addinsOutline_tex}()}.
}
