% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_addinsOutline_bookdown.R
\name{run_addinsOutline_Rmd_bookdown}
\alias{run_addinsOutline_Rmd_bookdown}
\title{Addins Displays Outline of an bookdown project and R Markdown file}
\usage{
run_addinsOutline_Rmd_bookdown()
}
\value{
RStudio Addins with browserViewer()
}
\description{
Addins Displays Outline of an bookdown project and R Markdown file with
and without child files.
}
\examples{
\dontrun{
run_addinsOutline_Rmd()
}
}
\seealso{
The function \code{\link{run_addinsOutline_tex}()}.
}
