% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_addinsOutline_tex.R
\name{run_addinsOutline_tex}
\alias{run_addinsOutline_tex}
\title{Addins Displays Outline of an tex file}
\usage{
run_addinsOutline_tex()
}
\value{
RStudio Addins with browserViewer()
}
\description{
Addins Displays Outline of an tex file with
and without child files insert with: \code{\\input{}} or
\code{\\include{}}.
}
\examples{
\dontrun{
run_addinsOutline_tex()
}
}
\seealso{
The function \code{\link{run_addinsOutline_Rmd}()} and \code{\link{run_addinsOutline_Rmd_bookdown}()}.
}
