% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_addinsOutline_Rmd.R
\name{run_addinsOutline_Rmd}
\alias{run_addinsOutline_Rmd}
\title{Addin for Displays Outline of an R Markdown Project}
\usage{
run_addinsOutline_Rmd()
}
\value{
'RStudio Addin' with browserViewer().
This 'RStudio' addin will show a list of the different sections established
in the R Markdown project and clicking on any element shown in the list
will cause 'RStudio' to show that section by opening the file
that contains it if it was not already open previously.
}
\description{
Addin for displays outline of an R Markdown project with
and without child files.
}
\examples{
if (interactive()) {
   library(addinsOutline)
   run_addinsOutline_Rmd()
}
}
\seealso{
The function \code{\link{run_addinsOutline_Rmd_bookdown}()} and
    \code{\link{run_addinsOutline_tex}()}.
}
