% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_addinsOutline_bookdown.R
\name{run_addinsOutline_Rmd_bookdown}
\alias{run_addinsOutline_Rmd_bookdown}
\title{Addin for Displays Outline of an Bookdown Project}
\usage{
run_addinsOutline_Rmd_bookdown()
}
\value{
RStudio Addin with browserViewer()
This 'RStudio' addin will show a list of the different sections established
in the 'bookdown' project (R Markdown files in directory project)
and clicking on any element shown in the list
will cause 'RStudio' to show that section by opening the file
that contains it if it was not already open previously.
To open a bookdown project select the file _bookdown.yml or index.Rmd.
}
\description{
Addin for displays outline of an bookdown project
(R Markdown files in directory project).
}
\examples{
if (interactive()) {
   library(addinsOutline)
   run_addinsOutline_Rmd_bookdown()
}
}
\seealso{
The function \code{\link{run_addinsOutline_tex}()} and
    \code{\link{run_addinsOutline_Rmd}()}.
}
